backend/requirements.txt
# ============================================
# ELDORET HOUSE HUNTERS - Backend Dependencies
# FastAPI Production Stack
# ============================================

# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
SQLAlchemy==2.0.35
pymysql==1.1.1
cryptography==44.0.0
alembic==1.13.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
bcrypt==4.2.1

# Image Processing
Pillow==11.0.0
python-magic==0.4.27

# Validation & Serialization
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

# CORS & Middleware
python-cors==1.0.0

# Utilities
python-dateutil==2.9.0.post0

# Development Tools (Optional)
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

